feat: add check that relationship entity exists

+++ app/services/cart_item/service.py
@@ -1,11 +1,11 @@
 from sqlalchemy.ext.asyncio.session import AsyncSession
 
-from core.models import CartItem
+from core.models import CartItem, Product
 from core.repositories.uow import UnitOfWork
 from services.cart_item.repository import CartItemRepository
 from services.cart_item.schemas import CartItemDTO, CartItemCreateSchema, CartItemUpdateSchema
 from services.product.schemas import ProductDTO
-from services.product.service import get_product_discount
+from services.product.service import get_product_discount, ProductNotFoundError
 
 
 class CartItemNotFoundError(Exception):
@@ -21,7 +21,7 @@ def get_cart_item_dto(item: CartItem) -> CartItemDTO:
     discount_info = get_product_discount(item.product)
     return CartItemDTO(
         total_price=discount_info.total_price,
-        discount=discount_info.discount_sumdiscount,
+        discount=discount_info.discount_sum,
         discount_description=discount_info.discount_description,
         price_with_discount=discount_info.price_with_discount,
         id=item.id,
@@ -60,25 +60,33 @@ class CartItemService:
         async with self.uow as uow:
             cart_item = await self.__find_by_id(uow.session, id)
             return get_cart_item_dto(cart_item)
+
     async def create(self, data: CartItemCreateSchema):
         async with self.uow as uow:
             await self.__validate_by_product_and_user(uow.session, data.user_id, data.product_id)
+            product = await uow.session.get(Product, data.product_id)
+            if not product:
+                raise ProductNotFoundError
             cart_item = await self.repository.create(uow.session, data.model_dump())
             await uow.commit()
             return get_cart_item_dto(cart_item)
 
-    async def update(self, data: CartItemUpdateSchema, id: int):
+    async def update(self, data: CartItemUpdateSchema, id: int, user_id: int) -> CartItemDTO:
         async with self.uow as uow:
             cart_item = await self.__find_by_id(uow.session, id)
+            if cart_item.user_id != user_id:
+                raise CartItemNotFoundError
             await self.repository.update(uow.session, data.model_dump(exclude_unset=True),
                                                              cart_item)
             await uow.commit()
             await uow.session.refresh(cart_item)
             return get_cart_item_dto(cart_item)
 
-    async def delete(self, id: int):
+    async def delete(self, id: int, user_id: int):
         async with self.uow as uow:
             cart_item = await self.__find_by_id(uow.session, id)
+            if cart_item.user_id != user_id:
+                raise CartItemNotFoundError
             await self.repository.delete(uow.session, cart_item)
             await uow.commit()
             return get_cart_item_dto(cart_item)
